Uso de JWT y reestructuración del proyecto
Se ha implementado el uso del estandar JWT para los usuarios y se ha cambiado la distribución del proyecto y su modelo.

+++ yourcourt/src/main/java/yourcourt/repository/FacilityRepository.java
@@ -12,7 +12,7 @@ import yourcourt.model.Facility;
 import yourcourt.model.FacilityType;
 
 @Repository
-public interface FacilityRepository extends CrudRepository<Facility, Integer>{
+public interface FacilityRepository extends CrudRepository<Facility, Long>{
 	
 	@Query("SELECT facilityTypes FROM FacilityType facilityTypes")
 	List<FacilityType> findFacilityTypes();
